+++ package.json
@@ -20,7 +20,6 @@
   "files": [
     "vaadin-*.d.ts",
     "vaadin-*.js",
-    "@types",
     "src",
     "theme"
   ],
