+++ vaadin-date-picker.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-date-picker.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
